requirements.txt
Django==1.5
psycopg2==2.5.1
wsgiref==0.1.2
django-widget-tweaks==1.3
django-eztables==0.3.2
